refactor UserInfo.kt into the firebase.database package
