+++ app/src/main/java/com/github/sdpcoachme/SignupActivity.kt
@@ -12,6 +12,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.testTag
 import com.github.sdpcoachme.database.Database
+import com.github.sdpcoachme.firebase.database.UserInfo
 import com.github.sdpcoachme.ui.theme.CoachMeTheme
 
 class SignupActivity : ComponentActivity() {
