--- app/src/main/java/com/github/sdpcoachme/UserInfo.kt
@@ -1,13 +1,0 @@
-package com.github.sdpcoachme
-
-/**
- * Data class for the client user
- */
-data class UserInfo(
-    val firstName: String,
-    val lastName: String,
-    val email: String,
-    val phone: String,
-    val location: String,
-    val sports: List<ListSport>
-)
